feat: rename top state
