+++ packages/axii-x6/examples/er2.jsx
@@ -6,9 +6,9 @@
 import { createElement, render, useRef } from "axii";
 import { debounce } from "lodash-es";
 
-import { EREditor2 } from "../";
+import { EREditor2 } from "../src/index";
 // FIXME: 使用 'axii-x6' 或者其他更合适的方式引入数据
-import localRawData from "../src/editors/er/data";
+import localRawData from "../src/editors2/er/data";
 
 const editorRef = useRef();
 const root = document.getElementById("root");
