+++ packages/axii-x6/src/k6/Edge.ts
@@ -3,7 +3,7 @@ import { ITopState, K6Node } from './Node';
 
 
 export abstract class K6Edge {
-  topState: ITopState;
+  data: ITopState;
   constructor (public target: K6Node) {
   }
 	// 同上，边线也通常是更简单的线，是否必要用Component渲染，或者默认模式
